requirements.txt
# Core data
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0

# HTTP / APIs
requests>=2.31.0
tenacity>=8.2.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.1.0
optuna>=3.4.0
shap>=0.43.0

# Statistical modelling
statsmodels>=0.14.0

# Experiment tracking
mlflow>=2.9.0

# Serialisation
joblib>=1.3.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Web application
streamlit>=1.28.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
